requirements.txt
pyyaml
round_robin_tournament
